Split numeric input example into Basic example and Extended example
