+++ packages/core/examples/index.ts
@@ -18,7 +18,8 @@ export * from "./focusExample";
 export * from "./hotkeyPiano";
 export * from "./hotkeyTester";
 export * from "./menuExample";
-export * from "./numericInputExample";
+export * from "./numericInputBasicExample";
+export * from "./numericInputExtendedExample";
 export * from "./nonIdealStateExample";
 export * from "./overlayExample";
 export * from "./popoverExample";
